// src/confluence-mcp-server/src/client.d.ts
export interface paths {
// ...
    "/pages": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get pages
         * @description Returns all pages. The number of results is limited by the `limit` parameter and additional results (if available)
         *     will be available through the `next` URL present in the `Link` response header.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to access the Confluence site ('Can use' global permission).
         *     Only pages that the user has permission to view will be returned.
         */
        get: operations["getPages"];
        put?: never;
        /**
         * Create page
         * @description Creates a page in the space.
         *
         *     Pages are created as published by default unless specified as a draft in the status field. If creating a published page, the title must be specified.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to view the corresponding space. Permission to create a page in the space.
         */
        post: operations["createPage"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/pages/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get page by id
         * @description Returns a specific page.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to view the page and its corresponding space.
         */
        get: operations["getPageById"];
        /**
         * Update page
         * @description Update a page by id.
         *
         *     When the "current" version is updated, the provided body content is considered as the latest version. This latest body content
         *     will be attempted to be merged into the draft version through a content reconciliation algorithm. If two versions are significantly diverged,
         *     the latest provided content may entirely override what was previously in the draft.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to view the page and its corresponding space. Permission to update pages in the space.
         */
        put: operations["updatePage"];
        post?: never;
        /**
         * Delete page
         * @description Delete a page by id.
         *
         *     By default this will delete pages that are non-drafts. To delete a page that is a draft, the endpoint must be called on a
         *     draft with the following param `draft=true`. Discarded drafts are not sent to the trash and are permanently deleted.
         *
         *     Deleting a page moves the page to the trash, where it can be restored later. To permanently delete a page (or "purge" it),
         *     the endpoint must be called on a **trashed** page with the following param `purge=true`.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to view the page and its corresponding space.
         *     Permission to delete pages in the space.
         *     Permission to administer the space (if attempting to purge).
         */
        delete: operations["deletePage"];
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/spaces": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get spaces
         * @description Returns all spaces. The results will be sorted by id ascending. The number of results is limited by the `limit` parameter and
         *     additional results (if available) will be available through the `next` URL present in the `Link` response header.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to access the Confluence site ('Can use' global permission).
         *     Only spaces that the user has permission to view will be returned.
         */
        get: operations["getSpaces"];
        put?: never;
        /**
         * Create space
         * @description Creates a Space as specified in the payload.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to create spaces.
         */
        post: operations["createSpace"];
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
    "/spaces/{id}/pages": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /**
         * Get pages in space
         * @description Returns all pages in a space. The number of results is limited by the `limit` parameter and additional results (if available)
         *     will be available through the `next` URL present in the `Link` response header.
         *
         *     **[Permissions](https://confluence.atlassian.com/x/_AozKw) required**:
         *     Permission to access the Confluence site ('Can use' global permission) and 'View' permission for the space.
         *     Only pages that the user has permission to view will be returned.
         */
        get: operations["getPagesInSpace"];
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
// ...
}
export type webhooks = Record<string, never>;
export interface components {
    schemas: {
        AdminKeyResponse: {
            /** @description User identifier. */
            accountId?: string;
            /**
             * Format: date-time
             * @description Timestamp in UTC that represents when the admin key will expire. In format "YYYY-MM-DDTHH:mm:ss.sssZ".
             */
            expirationTime?: string;
        };
        AttachmentSingle: {
            /** @description ID of the attachment. */
            id?: string;
// ...
            labels?: {
                results?: components["schemas"]["Label"][];
                meta?: components["schemas"]["OptionalFieldMeta"];
                _links?: components["schemas"]["OptionalFieldLinks"];
            };
// ...
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: {
        AdminKeyRequest: {
            content: {
                "application/json": {
                    /**
                     * Format: int32
                     * @description The requested duration of admin key access in minutes, up to a maximum of 60 minutes, after which the issued admin key will automatically expire.
                     */
                    durationInMinutes?: number;
                };
            };
        };
// ...
    };
    headers: never;
    pathItems: never;
}
export type $defs = Record<string, never>;
export interface operations {
    getAdminKey: {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        requestBody?: never;
        responses: {
            /** @description Returned if an admin key is currently enabled for the calling user. */
            200: {
                headers: {
                    [name: string]: unknown;
                };
                content: {
                    "application/json": components["schemas"]["AdminKeyResponse"];
                };
            };
            /** @description Returned if the authentication credentials are incorrect or missing from the request. */
            401: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
            /** @description Returned if the calling user does not currently have an admin key, if the calling user does not have permission to use admin keys, or if the site is not a Confluence Cloud Premium or Enterprise instance. */
            404: {
                headers: {
                    [name: string]: unknown;
                };
                content?: never;
            };
        };
    };
// ...
    getAttachments: {
// ...
        responses: {
            /** @description Returned if the requested attachments are returned. */
            200: {
// ...
                content: {
                    "application/json": {
                        results?: components["schemas"]["AttachmentBulk"][];
                        _links?: components["schemas"]["MultiEntityLinks"];
                    };
                };
            };
// ...
        };
    };
    getAttachmentById: {
        parameters: {
// ...
            path: {
                /** @description The ID of the attachment to be returned. If you don't know the attachment's ID, use Get attachments for page/blogpost/custom content. */
                id: string;
            };
            cookie?: never;
        };
// ...
    };
// ...
}